requirements.txt
langchain==0.0.18
PyPDF2==3.0.1
python-dotenv==0.21.0
streamlit==1.22.0
openai==0.27.6